Adding new sections (#36)
* Add required models

* Create GroupController

* Create SectionController and required models

* Create TeacherService and fix mapping

* Fix regex

* Fix Image data type from byte[] to string

* migrations fix

* Add AddressDto

* add DbSet<Section>

* Add required models

* Create GroupController

* Create SectionController and required models

* Create TeacherService and fix mapping

* remove migrations again

* update nuget packages

* Update branch after creating Repository

* Rename DTO models and fix namespaces

* Remove unused namespaces from ChildDTO

* Create OrganizationDTO and inject repository in SectionService

* Implement repositories for Section and Teacher services

* Fix Startup config for Section and Teacher services

* Enable XML documentation

* Fix merging error in AuthController

* Enable XML documentation in WebApi project

* rename tables and properties

* fix namespaces and delete useless folders "Implementation" and "ModelsDTO"

* add async to GetAllTeachers method & add keyword "this" in AuthController

* fix migrations after tables' renaming & remove own connectionstring

* rename all Section to Workshop

* move interfaces and their implementations to the "Service" folder

* create "Tools" folder and move "Mapping" into it

* add async to GetAllTeachers method

* fix WorkshopMapperProfile file & move it to the Tools folder

* modify Workshop service & controller

* modify TeacherService & controller

* modify OrganizationService, controller & models

* modify ChildService & controller

* fix datasourses

* Added Con Str to release config

* fix models and modelsDto

* move mapping folder inside the source folder

* create and apply mapping extensions

* renaming and namespace fixing

* add migration

* add newlines between properties

* rename ChildrenController to ChildController

* remove mapping profiles and extensions

* minor renaming

* create MappingExtension file

* update GetById method

+++ OutOfSchool/OutOfSchool.WebApi/Models/ChildDTO.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace OutOfSchool.WebApi.Models
 {
@@ -12,7 +9,9 @@ namespace OutOfSchool.WebApi.Models
         public string LastName { get; set; }
         public string MiddleName { get; set; }
         public DateTime DateOfBirth { get; set; }
-        public int ParentId { get; set; }
-        public int SocialGroupId { get; set; }
+        
+        public long ParentId { get; set; }
+        
+        public long SocialGroupId { get; set; }
     }
 }
